add find posts by user id

+++ Backend/src/main/java/com/websystique/springboot/service/PostService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface PostService {
 	List<Post> getAllPosts();
 
-	List<Post> getUserPosts(Long userId);
+	List<Post> getUserPosts(Integer userId);
 
 	void addPost(Post post);
 }
